requirements.txt
# dogeAgent Dependencies
# Core
langchain>=1.0.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-nvidia-ai-endpoints>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Voice
pygame>=2.5.0
pyaudio>=0.2.11
edge-tts>=6.1.0
pyttsx3>=2.90

# Weather & Search
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Encryption
cryptography>=41.0.0

# Dev
pytest>=7.4.0
pytest-asyncio>=0.21.0
